samples/microsoft/python/enterprise-agent-tutorial/2-prototype-to-production/requirements.txt
# Azure AI Foundry SDK v2.0 - Enterprise Agent Tutorial 2
# Prototype to Production

# Core Azure AI dependencies
azure-ai-projects>=2.0.0a20251015001
azure-ai-agents>=2.0.0a20251015001
azure-identity>=1.15.0

# Environment configuration
python-dotenv>=1.0.0

# JSON processing (built-in, but explicit for clarity)
# json - standard library

# Type hints support
typing-extensions>=4.9.0
